Move tflite examples to up dir
