+++ tank/albert_lite_base/input.json
