+++ tank/birds_V1/birds_V1_tflite_test.py
