+++ tank/cartoongan/cartoongan_tflite_test.py
