+++ tank/mobilenet_v1_224_1.0_uint8/mobilenet_v1_uint8_tflite_test.py
