+++ tank/mobilenet_v3.5multiavg_1.00_224_int8/mobilenet_v35_int8_tflite_test.py
