+++ tank/person_detect/person_detect_tflite_test.py
