+++ tank/ssd_mobilenet_v1_320_1.0_float/ssd_mobilenet_v1_float_tflite_test.py
