+++ tank/ssd_mobilenet_v1_320_1.0_uint8/ssd_mobilenet_v1_uint8_tflite_test.py
