+++ tank/ssd_mobilenet_v2_face_quant/ssd_mobilenet_v2_face_quant_tflite_test.py
