+++ tank/ssd_spaghettinet_edgetpu_large/ssd_spaghettinet_edgetpu_large_tflite_test.py
