+++ tank/ssd_spaghettinet_edgetpu_large_uint8/ssd_spaghettinet_edgetpu_large_uint8_tflite_test.py
